@@ -1,4 +1,6 @@
-import App from './App';
-import ReactDOM from 'react-dom';
+import App from "./App";
+import ReactDOM from "react-dom/client";
 
-ReactDOM.render(<App />, document.getElementById('root'));
+const root = ReactDOM.createRoot(document.getElementById("root"));
+
+root.render(<App />);
